requirements.txt
streamlit
pandas
scikit-learn
transformers
sentence-transformers
requests
cachetools
langchain
faiss-cpu
python-dotenv
